rename: change use color mode hook path
